Address comments

+++ flytekit/clients/auth/authenticator.py
@@ -9,7 +9,7 @@ import requests
 
 from . import token_client
 from .auth_client import AuthorizationClient
-from .exceptions import AccessTokenNotFoundError, AuthenticationError
+from .exceptions import AccessTokenNotFoundError, AuthenticationError, AuthenticationPending
 from .keyring import Credentials, KeyringStore
 
 
@@ -272,9 +272,6 @@ class DeviceCodeAuthenticator(Authenticator):
         session: typing.Optional[requests.Session] = None,
     ):
         cfg = cfg_store.get_client_config()
-        self._auth_client = None
-        self._authorization_endpoint = cfg.authorization_endpoint
-        self._redirect_uri = cfg.redirect_uri
         self._audience = audience or cfg.audience
         self._client_id = cfg.client_id
         self._device_auth_endpoint = cfg.device_authorization_endpoint
@@ -296,32 +293,29 @@ class DeviceCodeAuthenticator(Authenticator):
             verify=verify,
         )
 
-    def _initialize_auth_client(self):
-        if not self._auth_client:
-            self._set_header_key(self._header_key)
-            self._auth_client = AuthorizationClient(
-                endpoint=self._endpoint,
-                redirect_uri=self._redirect_uri,
-                client_id=self._client_id,
-                audience=self._audience,
-                scopes=self._scopes,
-                auth_endpoint=self._authorization_endpoint,
-                token_endpoint=self._token_endpoint,
-                verify=self._verify,
-                session=self._session,
-                refresh_access_token_params={},
-            )
-
     def refresh_credentials(self):
-        self._initialize_auth_client()
-        if self._creds:
+        if self._creds and self._creds.refresh_token:
             """We have an access token so lets try to refresh it"""
             try:
-                self._creds = self._auth_client.refresh_access_token(self._creds)
-                if self._creds:
-                    KeyringStore.store(self._creds)
+                access_token, refresh_token, expires_in = token_client.get_token(
+                    token_endpoint=self._token_endpoint,
+                    client_id=self._client_id,
+                    audience=self._audience,
+                    scopes=self._scopes,
+                    http_proxy_url=self._http_proxy_url,
+                    verify=self._verify,
+                    grant_type=token_client.GrantType.REFRESH_TOKEN,
+                    refresh_token=self._creds.refresh_token,
+                )
+                self._creds = Credentials(
+                    access_token=access_token,
+                    refresh_token=refresh_token,
+                    expires_in=expires_in,
+                    for_endpoint=self._endpoint,
+                )
+                KeyringStore.store(self._creds)
                 return
-            except AccessTokenNotFoundError:
+            except (AuthenticationError, AuthenticationPending):
                 logging.warning("Failed to refresh token. Kicking off a full authorization flow.")
                 KeyringStore.delete(self._endpoint)
